Update the approach of checking the CPUArchitecture
