+++ data/normalized_cpu_arch.csv
@@ -1,11 +1,12 @@
-Name,Architecture
-sandybridge,Sandy Bridge
-ivybridge,Ivy Bridge
-haswell,Haswell
-broadwell,Broadwell
-skylake,Sky Lake
-cascadelake,Cascade Lake
-coffeelake,Coffee Lake
-alderlake,Alder Lake
-cascadelake,Cascade Lake
-neoverse_n1,neoverse_n1
+Architecture
+Sandy Bridge
+Ivy Bridge
+Haswell
+Broadwell
+Sky Lake
+Cascade Lake
+Coffee Lake
+Alder Lake
+Cascade Lake
+Sapphire Rapids
+neoverse_n1
